refactor task models
non db related models moved to common
